requirements.txt
# Google ADK runtime
google-adk>=1.0.0

# Async PostgreSQL driver
asyncpg>=0.29.0

# Environment variable loading
python-dotenv>=1.0.0
